make compareCourse and compareCourseAgainstQualification consistent
compareCourse -> compareCourseToCourse
compareCourseAgainstQualification -> compareCourseToQualification
