+++ lib/check-for-course.js
@@ -1,5 +1,5 @@
 import some from 'lodash/collection/some'
-import compareCourse from './compare-course'
+import compareCourseToCourse from './compare-course-to-course'
 
 /**
  * Checks if a course exists in an array of courses
@@ -9,5 +9,5 @@ import compareCourse from './compare-course'
  * @returns {Boolean} - if the course was found or not
  */
 export default function checkForCourse(query, courses) {
-    return some(courses, (c) => compareCourse(query, c))
+    return some(courses, (c) => compareCourseToCourse(query, c))
 }
