+++ lib/compare-course-to-course.js
@@ -22,7 +22,7 @@ const baseKeys = [
  * @param {Course} other - the course to compare against
  * @returns {boolean} - if the course matched
  */
-export default function compareCourse(query, other) {
+export default function compareCourseToCourse(query, other) {
     // If the query is more specific than the one being compared to, and
     // things don't match, return false.
     // But, if the query is *less* specific than the other course, only check
