+++ lib/compare-course-to-qualification.js
@@ -12,9 +12,9 @@ import assertKeys from './assert-keys'
  * @param {string} $value - the value compare to
  * @returns {boolean} - whether the course matched or not
  */
-export default function compareCourseAgainstQualification(course, {$key, $operator, $value}) {
+export default function compareCourseToQualification(course, {$key, $operator, $value}) {
     if (isArray($value)) {
-        throw new TypeError(`compareCourseAgainstQualification(): what would a comparison to a list even do? oh, wait; i suppose it could compare against one of several values... well, I'm not doing that right now. if you want it, edit the PEG and stick appropriate stuff in here (probably simplest to just call this function again with each possible value and return true if any are true.)`)
+        throw new TypeError(`compareCourseToQualification(): what would a comparison to a list even do? oh, wait; i suppose it could compare against one of several values... well, I'm not doing that right now. if you want it, edit the PEG and stick appropriate stuff in here (probably simplest to just call this function again with each possible value and return true if any are true.)`)
     }
 
     else if (isObject($value)) {
@@ -22,7 +22,7 @@ export default function compareCourseAgainstQualification(course, {$key, $operat
         // operators earlier, in the filterByQualification function.
         assertKeys($value, '$computed-value')
         const simplifiedOperator = {$key, $operator, $value: $value['$computed-value']}
-        return compareCourseAgainstQualification(course, simplifiedOperator)
+        return compareCourseToQualification(course, simplifiedOperator)
     }
 
     else {
