+++ lib/filter-by-where-clause.js
@@ -1,5 +1,5 @@
 import assertKeys from './assert-keys'
-import compareCourseAgainstQualification from './compare-course-against-qualification'
+import compareCourseToQualification from './compare-course-to-qualification'
 import filter from 'lodash/collection/filter'
 import flatten from 'lodash/array/flatten'
 import forEach from 'lodash/collection/forEach'
@@ -96,7 +96,7 @@ export function filterByQualification(list, qualification, fullList) {
     }
 
     const filtered = filter(list,
-        course => compareCourseAgainstQualification(course, qualification))
+        course => compareCourseToQualification(course, qualification))
 
     return filtered
 }
