+++ lib/find-course.js
@@ -1,5 +1,5 @@
 import find from 'lodash/collection/find'
-import compareCourse from './compare-course'
+import compareCourseToCourse from './compare-course-to-course'
 
 /**
  * Finds a course in a list of courses
@@ -9,5 +9,5 @@ import compareCourse from './compare-course'
  * @returns {Course|undefined} - the found course
  */
 export default function findCourse(query, courses) {
-    return find(courses, (c) => compareCourse(query, c))
+    return find(courses, (c) => compareCourseToCourse(query, c))
 }
