+++ lib/get-occurrences.js
@@ -3,7 +3,7 @@ import simplifyCourse from './simplify-course'
 
 // old version; compares course objects instead of simplified versions
 // export default function getOccurrences(course, courses) {
-//     return filter(courses, (c) => compareCourse(filter, c))
+//     return filter(courses, (c) => compareCourseToCourse(filter, c))
 // }
 
 /**
